/** REFACTOR **/
Refactored the error message that are received by create and edit views by creating a seperate errors.php file in partials and modified loadPartial helper to accept an array as second parameter so that the error messages can be used on a different view.

Created a post route to register a user and store() method inside the UserController to add validations in the form and later save the user to the database.